User Authentication and Authorization added

+++ backend/routes/posts.js
@@ -2,6 +2,7 @@ const express = require("express");
 const multer = require("multer");
 
 const Post = require("../models/post");
+const checkAuth = require("../middleware/check-auth");
 
 const router = express.Router();
 
@@ -30,15 +31,18 @@ const storage = multer.diskStorage({
   }
 });
 
+// AUTHENTICATED
 router.post(
   "", 
+  checkAuth,
   multer({storage: storage}).single("image"), 
   (req, res, next) => {
     const url = req.protocol + '://' + req.get("host");
     const post = new Post({
       title: req.body.title,
       content: req.body.content,
-      imagePath: url + "/images/" + req.file.filename
+      imagePath: url + "/images/" + req.file.filename,
+      creator: req.userData.userId
     });
     post.save().then(createdPost => {
       res.status(201).json({
@@ -52,8 +56,10 @@ router.post(
   }
 );
 
+// AUTHENTICATED
 router.put(
-  "/:id", 
+  "/:id",
+  checkAuth, 
   multer({ storage: storage}).single("image"),
   (req, res, next) => {
     let imagePath = req.body.imagePath;
@@ -65,15 +71,20 @@ router.put(
       _id: req.body.id,
       title: req.body.title,
       content: req.body.content,
-      imagePath: imagePath
+      imagePath: imagePath,
+      creator: req.userData.userId
     });
-    console.log(post);
-    Post.updateOne({ _id: req.params.id }, post).then(result => {
-      res.status(200).json({ message: "Update successful!" });
+    Post.updateOne({ _id: req.params.id, creator: req.userData.userId }, post).then(result => {
+      if (result.nModified > 0) {
+        res.status(200).json({ message: "Update successful!" });
+      } else {
+        res.status(401).json({ message: "Not Authorized" })
+      }
     });
   }
 );
 
+// OPEN
 router.get("", (req, res, next) => {
   const pageSize = +req.query.pagesize;
   const currentPage = +req.query.page;
@@ -97,6 +108,7 @@ router.get("", (req, res, next) => {
   });
 });
 
+// OPEN
 router.get("/:id", (req, res, next) => {
   Post.findById(req.params.id).then(post => {
     if (post) {
@@ -107,10 +119,14 @@ router.get("/:id", (req, res, next) => {
   });
 });
 
-router.delete("/:id", (req, res, next) => {
-  Post.deleteOne({ _id: req.params.id }).then(result => {
-    console.log(result);
-    res.status(200).json({ message: "Post deleted!" });
+// AUTHENTICATED
+router.delete("/:id", checkAuth, (req, res, next) => {
+  Post.deleteOne({ _id: req.params.id, creator: req.userData.userId }).then(result => {
+    if (result.n > 0) {
+      res.status(200).json({ message: "Post deleted!" });
+    } else {
+      res.status(401).json({ message: "Not Authorized" })
+    }
   });
 });
